+++ js/util.js
@@ -1,0 +1,13 @@
+//Модуль с вспомогательными функциями
+
+//Функция для получения случайного числа
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+//Функция для получения элементат массива по индексу
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export { getRandomArrayElement, getRandomInteger };
